+++ packages/kofono/src/index.ts
@@ -1,23 +1,36 @@
+// builder
 export { Builder } from "./builder/Builder";
 export { buildSchema, tryBuildSchema } from "./builder/helpers";
 export { K } from "./builder/K";
 export { SchemaBuilder } from "./builder/SchemaBuilder";
 export * from "./builder/types";
+// common
 export { isObjectLiteral } from "./common/helpers";
+// extension
 export { ExtensionsFactory } from "./extension/ExtensionsFactory";
+// form
 export { defaultConfig, defaultPassHandler } from "./form/defaults";
 export * from "./form/events/types";
 export { Form } from "./form/Form";
 export { FormComponent } from "./form/FormComponent";
 export { FormProperty } from "./form/FormProperty";
 export * from "./form/types";
+// property
 export * from "./property/categories";
 export { Property } from "./property/Property";
 export * from "./property/types";
+// schema
 export * from "./schema/Schema";
-export { DataSelector } from "./selector/DataSelector";
+// selector
+export {
+    DataSelector,
+    DataSelectorIndexOutOfBoundsError,
+    DataSelectorNotFoundError,
+} from "./selector/DataSelector";
 export { GenericDataQuerier } from "./selector/GenericDataQuerier";
 export { ReadOnlyDataSelector } from "./selector/ReadOnlyDataSelector";
+export * from "./selector/types";
+// validator
 export * from "./validator/builtinValidators";
 export * from "./validator/schema";
 export * from "./validator/types";